fixed ftnchek warnings for mc_exact.f

+++ mc_adapt.f
@@ -88,11 +88,11 @@ CCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCC
 
       real*8 r_samp, c
 
-      c = - (k_max * 1.0/V_comp / log(1 - p_max))
+      c = - (k_max * 1d0/V_comp / log(1 - p_max))
       del_t = r_samp_max / c
       if (del_t .gt. del_t_max) del_t = del_t_max
       r_samp = del_t * c
-      n_samp = r_samp * M*(M-1)/2
+      n_samp = int(r_samp * M*(M-1)/2)
 
       return
       end
